feat: 메인 타입 분리
