+++ src/types/index.ts
@@ -1,6 +1,7 @@
 /* 메인 페이지 */
 export * from './main/recommend';
 export * from './main/sectionTitle';
+export * from './main/popular';
 export * from './main/recentArt';
 export * from './main/exhibition';
