+++ src/types/main/popular.ts
@@ -1,0 +1,10 @@
+export interface Exhibition {
+  id: string;
+  title: string;
+  location: string;
+  imageUrl: string;
+}
+
+export interface PopularExhibitionSectionProps {
+  exhibitions: Exhibition[];
+}
